refactor profile page routing
